sensu_check_config -> sensu_check, add sensu_check_config for config, all params in sensu::check
